backend/requirements.txt
Flask
torch
torchvision
Pillow
numpy
mediapipe
flask-cors